Moving Crypto class into Helpers - #271
